picassowifi: Hexedit /vendor/lib/egl/libGLES_mali.so
The L linker changed dlsym to do a breadth-first search for symbols
rather than looking only in the referenced library.

Our mali blob is linked against libgui.so which is linked against
libEGL.so which causes all kinds of problems because vendor
implementations get resolved to the opengl generic wrappers.

Hexedit this problem away by changing libgui.so to libui.so (which
it is already linked against and therefore safe to link against
again).  Apparently they didn't really need to link against
libgui.so?

Change-Id: I5ac7ac5c0b145cdbbc4930170d46038d634e8d1c